Change PowerParam.value from float to int across AST, parser, serializer, and spec
Mirrors the earlier RPE change — power in watts is always whole numbers.
Parser rounds fractional input, resolver rounds computed values.
Simplifies serializer/cli/Grit display code that checked int-vs-float.

+++ src/owf/ast/params.py
@@ -29,7 +29,7 @@ class PercentOfParam:
 class PowerParam:
     """Literal power in watts, e.g. @200W."""
 
-    value: float
+    value: int
     span: SourceSpan | None = field(default=None, compare=False, repr=False)
